Manager: Remove unused variables
From PVS Studio:
V808
'strBuffer' object of 'wxString' type was created but was not utilized.
'strValue' object of 'wxString' type was created but was not utilized.
'strName' object of 'wxString' type was created but was not utilized.
'strURL' object of 'wxString' type was created but was not utilized.
'strTeamName' object of 'wxString' type was created but was not utilized.
'strDialogMessage' object of 'wxString' type was created but was not utilized.
'strOldLanguageCode' object of 'wxString' type was created but was not utilized.
'strHiddenColumns' object of 'wxString' type was created but was not utilized.
'strConnectedCompter' object of 'wxString' type was created but was not utilized.
'strEmpty' object of 'wxString' type was created but was not utilized.
'value_w' object of 'basic_string' type was created but was not utilized.
'pct_done_str' object of 'wxString' type was created but was not utilized.
'bmpPrev' object of 'wxBitmap' type was created but was not utilized.
https://www.viva64.com/en/w/V808/print

Signed-off-by: Vitalii Koshura <lestat.de.lionkur@gmail.com>